celelytics-admin-portal/requirements.txt
Flask
python-dotenv
azure-data-tables
azure-core
pytz
